+++ src/compas/files/off.py
@@ -2,13 +2,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-try:
-    from urllib.request import urlopen
-except ImportError:
-    from urllib2 import urlopen
-
-from compas.utilities import geometric_key
-
+from compas.files.BaseReader import BaseReader
 
 __all__ = [
     'OFF',
@@ -17,7 +11,11 @@ __all__ = [
 
 
 class OFF(object):
-    """Read and write files in OFF format.
+    """Read and write files in *OFF* format.
+
+    Notes
+    -----
+    GeomView Object File Format
 
     References
     ----------
@@ -25,74 +23,44 @@ class OFF(object):
     * http://www.geomview.org/docs/html/OFF.html
     * http://segeval.cs.princeton.edu/public/off_format.html
 
-
     """
     def __init__(self, filepath):
         self.reader = OFFReader(filepath)
 
 
-class OFFReader(object):
-    """Read the contents of an *obj* file.
+class OFFReader(BaseReader):
+    """Read the contents of an *off* file.
 
     Parameters
     ----------
-    filepath : str
-        Path to the file.
+    location: str or Path object
+        Path or URL to the file.
 
     Attributes
     ----------
     vertices : list
         Vertex coordinates.
-    weights : list
-        Vertex weights.
-    textures : list
-        Vertex textures.
-    normals : list
-        Vertex normals.
-    points : list
-        Point objects, referencing the list of vertices.
-    lines : list
-        Line objects, referencing the list of vertices.
     faces : list
         Face objects, referencing the list of vertices.
-    curves : list
-        Curves
-    curves2 : list
-        Curves
-    surfaces : list
-        Surfaces
-
-    Notes
-    -----
-    For more info, see [1]_.
-
-    References
-    ----------
-    .. [1] Bourke, P. *Object Files*.
-           Available at: http://paulbourke.net/dataformats/obj/.
+    vertex_count : int
+        Vertex count stated in beginning of file
+    face_count : int
+        Face count stated in beginning of file
+    edge_count : int
+        Edge count stated in beginning of file
 
     """
-
-    def __init__(self, filepath):
-        self.filepath = filepath
-        self.content = None
+    def __init__(self, location):
+        super(OFFReader, self).__init__(location)
+        self.content = self.read_from_location()
         self.vertices = []
         self.faces = []
-        self.v = 0
-        self.f = 0
-        self.e = 0
-        self.open()
+        self.vertex_count = 0
+        self.face_count = 0
+        self.edge_count = 0
         self.pre()
         self.read()
         self.post()
-
-    def open(self):
-        if self.filepath.startswith('http'):
-            resp = urlopen(self.filepath)
-            self.content = iter(resp.read().decode('utf-8').split('\n'))
-        else:
-            with open(self.filepath, 'r') as fh:
-                self.content = iter(fh.readlines())
 
     def pre(self):
         lines = []
@@ -128,11 +96,11 @@ class OFFReader(object):
 
         """
         if not self.content:
-            return
+            raise RuntimeError('Import failed')
 
         header = next(self.content)
-        if not header.lower() == 'off':
-            return
+        if header.lower() != 'off':
+            raise RuntimeError('Not a valid OFF file')
 
         for line in self.content:
             if line.startswith('#'):
@@ -143,7 +111,8 @@ class OFFReader(object):
                 continue
 
             if len(parts) == 3:
-                self.v, self.f, self.e = int(parts[0]), int(parts[1]), int(parts[2])
+                self.vertex_count, self.face_count, self.edge_count = \
+                    int(parts[0]), int(parts[1]), int(parts[2])
                 break
 
         for line in self.content:
@@ -160,7 +129,6 @@ class OFFReader(object):
                 if f == len(parts[1:]):
                     self.faces.append([int(index) for index in parts[1:]])
                 continue
-
 
 
 # ==============================================================================